Refactor ApplePayController to fit the flow
